requirements.txt
ansible==10.3.0
ansible-builder==3.1.0
ansible-compat==24.7.0
ansible-core==2.17.3
ansible-creator==24.7.1
ansible-dev-environment==24.7.0
ansible-dev-tools==24.7.2
ansible-lint==24.7.0
ansible-navigator==24.8.0
ansible-runner==2.4.0
ansible-sign==0.1.1
apturl==0.5.2
attrs==24.2.0
bindep==2.11.0
black==24.8.0
blinker==1.4
bracex==2.5
Brlapi==0.8.3
cachetools==5.5.0
certifi==2020.6.20
cffi==1.17.0
chardet==5.2.0
click==8.0.3
click-help-colors==0.9.4
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.21.2
enrich==1.2.7
exceptiongroup==1.2.2
execnet==2.1.1
filelock==3.15.4
google-auth==2.34.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kubernetes==30.1.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
molecule==24.8.0
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
onigurumacffi==1.3.0
packaging==24.1
Parsley==1.3
passlib==1.7.4
pathspec==0.12.1
pbr==6.0.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
protobuf==3.12.4
ptyprocess==0.7.0
pyaml==24.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject-api==1.7.1
pyRFC3339==1.1
pytest==8.3.2
pytest-ansible==24.8.0
pytest-xdist==3.6.1
python-apt==2.4.0+ubuntu3
python-daemon==3.0.1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.5.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.35.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
resolvelib==1.0.1
rich==13.7.1
robotframework==7.0.1
robotframework-jsonlibrary==0.5
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
subprocess-tee==0.4.2
systemd-python==234
tomli==2.0.1
tox==4.18.0
tox-ansible==24.8.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
vboxapi==1.0
virtualenv==20.26.3
wadllib==1.3.6
wcmatch==9.0
websocket-client==1.8.0
xdg==5
xkit==0.0.0
yamllint==1.35.1
zipp==1.0.0
